Use the absolute path for the tag release python script
The workflow for generating a new tag and release referenced a python
script for getting the new tag via its relative path. This caused
issues when the workflow was used in other repos as the relative path
within the importing repo was used instead of the relative path within
this repo.

This commit fixes this by using the absolute GitHub path of the python
script.