Use relative path for the create_main_release workflow script and readme updates
Using an absolute path to the script did not fix the issue with using the
create_main_release workflow in other repos, so this commit reverts it to the
correct relative path.

Updates to the README with instructions for using a workflow from this repo in
your own repo and a disclaimer about the create_main_release workflow.