Add an action for auto-assigning PRs to the developers group
We have a developers group that has all of the engineering GH accounts.

We've historically used this in order to make sure that PRs appear
in the list of PRs to be reviewed for all developers, however, we've
previously had to do this manually and we'd sometimes forget to do this.

This PR adds an action that automatically assigns all PRs to the developers
group.